+++ common/src/main/java/com/selfdot/cobblemontrainers/menu/SetupMenu.java
@@ -1,4 +1,4 @@
-package com.selfdot.cobblemontrainers;
+package com.selfdot.cobblemontrainers.menu;
 
 import com.cobblemon.mod.common.CobblemonItems;
 import com.cobblemon.mod.common.api.abilities.AbilityTemplate;
@@ -11,6 +11,7 @@ import com.cobblemon.mod.common.api.pokemon.stats.Stats;
 import com.cobblemon.mod.common.battles.pokemon.BattlePokemon;
 import com.cobblemon.mod.common.pokemon.Pokemon;
 import com.cobblemon.mod.common.pokemon.Species;
+import com.selfdot.cobblemontrainers.CobblemonTrainers;
 import com.selfdot.cobblemontrainers.trainer.Trainer;
 import com.selfdot.cobblemontrainers.trainer.TrainerPokemon;
 import com.selfdot.libs.cobblemon.CobblemonUtils;